feat: add key derivation flow
during first boot, if there is no salt or key
in the board, the key derivation creation flow
will start. During this flow, the User has to
enter a password with length between 8 and 12 characters,
which is then used to generate a key.
This key is stored in SPIFFS, and it is  used
to encrypt and decrypt services secrets during boot.